Auth Tutorial Added

+++ student_app/views.py
@@ -1,11 +1,24 @@
-from django.shortcuts import render, HttpResponse
+from django.shortcuts import render, HttpResponse, redirect
 from .models import Student
-from .form import StudentForm
+from .form import StudentForm, RegisterForm
 
 
 # Create your views here.
 def home(request):
     return render(request, "home.html")
+
+def profile(request):
+    if not request.user.is_authenticated:
+        return redirect("login")
+    return render(request, "accounts/profile.html")
+
+
+def register(request):
+    form = RegisterForm(request.POST)
+    if form.is_valid():
+        form.save()
+        return HttpResponse("<h1>Registration Successful</h1>")
+    return render(request, "registration/register.html", {"form": form})
 
 
 # view all
@@ -14,18 +27,22 @@ def students(request):
         form = StudentForm(request.POST)
 
         if request.method == "GET":
-            students = Student.objects.all()
-            return render(request, "students.html", {"students": students, "form": form})
+            students = request.user.students.all()
+            return render(
+                request, "students.html", {"students": students, "form": form}
+            )
 
         # django form
         if request.method == "POST":
             if form.is_valid():
                 form.save()
+                request.user.students.add(form.instance)
                 return HttpResponse("<h1>Upload Successful</h1>")
 
         # regular form
         # if request.method == "POST":
         #     student = Student()
+        #     student.owner = request.POST.get(request.user)
         #     student.name = request.POST.get("name")
         #     student.email = request.POST.get("email")
         #     student.about = request.POST.get("about")
@@ -33,10 +50,9 @@ def students(request):
 
         #     student.save()
         #     return HttpResponse("<h1>Upload Successful</h1>")
-        
+
     except:
         return HttpResponse("<h1>Upload Failed</h1>")
-        
 
 
 def student(request, pk):
@@ -45,14 +61,19 @@ def student(request, pk):
         form = StudentForm(request.POST, instance=student)
 
         if request.method == "GET":
-            return render(request, "student.html", {"student": student, "form": form})
+            if request.user.username == student.owner.username:
+                return render(request, "student.html", {"student": student, "form": form})
+            else:
+                return HttpResponse("<h1>Unauthorized</h1>")
 
         # django
         if request.method == "POST":
-            if form.is_valid():
-                form.save()
-                return HttpResponse("<h1>Upload Successful</h1>")
-
+            if request.user.username == student.owner.username:
+                if form.is_valid():
+                    form.save()
+                    return HttpResponse("<h1>Upload Successful</h1>")
+            else:
+                return HttpResponse("<h1>Unauthorized</h1>")
 
         # html
         # if request.method == "POST":
@@ -68,9 +89,13 @@ def student(request, pk):
 
 
 def student_delete(request, pk):
-    try:
+    try:        
         student = Student.objects.get(pk=pk)
-        student.delete()
-        return HttpResponse("<h1>Student Deleted Successfully</h1>")
+        
+        if request.user.username == student.owner.username:
+            student.delete()
+            return HttpResponse("<h1>Student Deleted Successfully</h1>")
+        else:
+            return HttpResponse("<h1>Unauthorized</h1>")
     except:
         return HttpResponse("<h1>Delete Failed</h1>")
